+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './api';
 export * from './frame';
 export * from './logger';
 export * from './notifications';
+export * from './tests';